Fix broken entity decompression
As identified by @fabn in #13, the `dump` logic for compressed entities
is broken for a wide set of values. From the issue:

> When compressed data size is below the compression threshold data
> loading is completely wrong. It will return the compressed data.

This commit mitigates the problem by introducing a `compressed?`
predicate that inspects the initial bytes of the value and tries to
determine if it is in fact compressed. There is a possibility of false
negatives for compressed values, in which case the decompression attempt
will be rescued and return the initial value.

Overall the impact of this change is very minor, in a microbenchmark the
corrected loading behavior is `1.04x` slower.

```
Calculating -------------------------------------
         entity:load     4.496k i/100ms
 entity:load_correct     4.257k i/100ms
-------------------------------------------------
         entity:load     58.229k (± 8.0%) i/s -    292.240k
 entity:load_correct     55.807k (± 2.1%) i/s -    280.962k

Comparison:
         entity:load:    58228.7 i/s
 entity:load_correct:    55806.8 i/s - 1.04x slower
```