Allow dumping and loading nil values
According to the caching guides for active support it is valid to cache
`nil` values. Previously that was possible with readthis, but it
wouldn't marshal the `nil` value, which forced it into an empty string.
Later, when the blank string was read back it couldn't be loaded,
instead raising `ArgumentError: marshal data too short`.

Closes #15