Move to a binary flag format for entities
Inspired by @fabn's proof of concept, this switches from a lengthy ASCII
based marker system to a single byte for flags.

The upside is that parsing is quite a bit faster and much less space is
used for each entity. The downside is that a single byte only allows for
up to 8 different marshalers, all of which must be hard coded. This
limits the flexibility that was possible before, but is worth the trade off
because of significant wins in consistency, allocation and entity size.