Handle case when nil is explicitly passed to fetch
I ran in to this issue while using the active model serializers gem
version v0.10.0.rc3.

They default their options that get passed to the cache as nil if
none are set. Using an options hash doesn't guard against explicitly
passed parameters, in this case nil. That caused an error, shown in
the screenshot below.

I hope this helps :)