Reject nil values from read_multi output
Previously `nil` values were always preserved when fetching with
`read_multi`. This differed from the `ActiveSupport::Cache` base
behaviour, and causes bugs in Rails 5.0. Because this is a breaking
change, and is required for the proper operation of `fetch_multi`, it
is configurable.

The old fetching behaviour, with `nil` intact, can be achieved by
passing the `retain_nils` flag:

`cache.read_multi('a', 'b', retain_nils: true)`

This can also be set as a global option to retain backward
compatibility.