Prevent encoding issues when using UTF-8 keys
Setting both the key and value to UTF-8 strings caused differing
encodings to be sent to Redis. The key was UTF-8 but the value would be
ASCII-8BIT, because it was compressed and then packed as 8bit integers.
The workaround is to force the encoding of both strings to binary,
something that will always be compatible.

Closes #52